backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
beautifulsoup4
requests
python-dotenv
langchain-google-genai
langchain
pydantic